+++ CHANGELOG.md
@@ -14,7 +14,8 @@
 
 #### Changes
 
-* Renamed `Status::PayloadTooLarge` to `Status::ContentTooLarge`
+* Renamed `Status::PayloadTooLarge` to `Status::ContentTooLarge`.
+* The input validation middleware will now default to status `422` (unprocessable entity) instead of `400` (bad request) for invalid requests.
 
 --------------------------------------------------------